refactor(bookmark): package 구조 변경
