+++ backend/core/src/main/java/com/wooacourse/taggle/bookmark/exception/EmptyValueException.java
@@ -1,4 +1,4 @@
-package com.wooacourse.taggle.domain.bookmark.exception;
+package com.wooacourse.taggle.bookmark.exception;
 
 public class EmptyValueException extends RuntimeException {
     public EmptyValueException(String message) {
