+++ backend/core/src/main/java/com/wooacourse/taggle/bookmark/exception/UrlFormatMismatchException.java
@@ -1,4 +1,4 @@
-package com.wooacourse.taggle.domain.bookmark.exception;
+package com.wooacourse.taggle.bookmark.exception;
 
 public class UrlFormatMismatchException extends RuntimeException {
     public UrlFormatMismatchException(String message) {
